+++ client/src/features/admin-competition/model/service.ts
